agents-service/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
supabase==2.28.3
pydantic-settings==2.6.1
PyJWT==2.12.0
slowapi==0.1.9
httpx==0.27.2
apscheduler==3.10.4
anthropic>=0.40.0
langgraph>=0.2.0
langgraph-cli>=0.1.55
langchain-ollama>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langsmith>=0.1.0
docker>=7.0.0
PyGithub>=2.1.0
psycopg2-binary>=2.9.9
tenacity>=8.2.3
prometheus-client>=0.20.0
